requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
claude-agent-sdk==0.1.68
click==8.3.3
cryptography==47.0.0
duckdb==1.5.2
fastapi==0.136.1
-e git+https://github.com/mehdidotesk73/forge-framework.git@0a908f0e2220f33f601c3f2d64235fdabecb2038#egg=forge_framework&subdirectory=packages/forge-py
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pyarrow==24.0.0
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
referencing==0.37.0
rich==15.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sse-starlette==3.4.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
